% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRM_SCREAM_calculate_colony_density.R
\name{DRM_SCREAM_calculate_colony_density}
\alias{DRM_SCREAM_calculate_colony_density}
\title{Creates colony density and colony size summary dataframes}
\usage{
DRM_SCREAM_calculate_colony_density(project = "NULL", species_filter = "NULL")
}
\arguments{
\item{project}{A string indicating the project, DRM or DRM and SCREAM combined}

\item{species_filter}{A boolean value indicating whether to filter to a subset of species}
}
\value{
A dataframe
}
\description{
Creates colony density and colony size summary dataframes
}
