% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRM_SCREAM_calculate_disease_prevalence.R
\name{DRM_SCREAM_calculate_disease_prevalence}
\alias{DRM_SCREAM_calculate_disease_prevalence}
\title{Creates species list, species richness and species diversity dataframes from NCRMP and FRRP data}
\usage{
DRM_SCREAM_calculate_disease_prevalence(
  project = "NULL",
  species_filter = "NULL"
)
}
\arguments{
\item{project}{A string indicating the project, DRM or DRM and SCREAM combined}

\item{species_filter}{A boolean value indicating whether to filter to a subset of species}
}
\value{
A dataframe
}
\description{
Creates species list, species richness and species diversity dataframes from NCRMP and FRRP data
}
