% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/DRM_SCREAM_calculate_species_richness_diversity.R
\name{DRM_SCREAM_calculate_species_richness_diversity}
\alias{DRM_SCREAM_calculate_species_richness_diversity}
\title{Creates species list, species richness and species diversity dataframes from NCRMP and FRRP data}
\usage{
DRM_SCREAM_calculate_species_richness_diversity()
}
\value{
A dataframe
}
\description{
Creates species list, species richness and species diversity dataframes from NCRMP and FRRP data
}
