% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRM_SCREAM_make_weighted_demo_data.R
\name{DRM_SCREAM_make_weighted_demo_data}
\alias{DRM_SCREAM_make_weighted_demo_data}
\title{Creates weighted demo data}
\usage{
DRM_SCREAM_make_weighted_demo_data(inputdata, datatype)
}
\arguments{
\item{inputdata}{A dataframe}

\item{datatype}{A string indicating the datatype}
}
\value{
A dataframe
}
\description{
Creates weighted demo data
}
