% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_grid_data.R
\name{extract_grid_data}
\alias{extract_grid_data}
\title{Overlay site coordinates on a grid and extract grid info}
\usage{
extract_grid_data(field_sites, grid, region)
}
\arguments{
\item{grid}{A spatialPolygonsDataFrame}

\item{demo_sites}{A dataframe}
}
\value{
A spatialPolygonsDataFrame
}
\description{
This function imports .RData files with FL grids, reprojects the grids to WGS84, subsets coordinates from the data, projects those coordinates to WGS84, extracts grid data using the subset coordinates,
and merges grid data with the coordinates back to the original dataframe. The function outputs demo sites with grid information for  FLK, DT, and SEFCRI.
}
\concept{grid extractions}
