% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_grids.R
\name{import_grids}
\alias{import_grids}
\title{Import NCRMP grids and save as R data}
\usage{
import_grids(region, year)
}
\arguments{
\item{region}{A string indicating the region}

\item{year}{A numeric indicating the year}
}
\value{
A list of two dataframes: sampling_grid and strata_ntot
}
\description{
This function has 4 purposes:
 (1) Import sample frame grids for various regions;
 (2) Reproject from UTM to WGS 1984;
 (3) Save as R.data for use in extract grid function;
 (4) Calculate Ntot and output to a csv file
}
