% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_NCRMP_benthic_cover_data.R
\name{load_NCRMP_benthic_cover_data}
\alias{load_NCRMP_benthic_cover_data}
\title{Creates combined cover dataframes}
\usage{
load_NCRMP_benthic_cover_data(project = "NULL", region)
}
\arguments{
\item{project}{A string indicating the project: "NCRMP" or "MIR". Default is NCRMP.}

\item{region}{A string indicating the region}
}
\value{
A dataframe
}
\description{
Creates combined cover dataframes
}
