% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_NCRMP_DRM_demo_data.R
\name{load_NCRMP_DRM_demo_data}
\alias{load_NCRMP_DRM_demo_data}
\title{Creates combined demo dataframes}
\usage{
load_NCRMP_DRM_demo_data(project = "NULL", region, species_filter = "NULL")
}
\arguments{
\item{project}{A string indicating the project, NCRMP, NCRMP and DRM combined, or MIR. Default is NCRMP.}

\item{region}{A string indicating the region. Options are: "SEFCRI", "FLK", "Tortugas", "STX", "STTSTJ", "PRICO", and "GOM".}

\item{species_filter}{An optional string indicating the species to subset to, if desired. Format is the 7 digits species code (e.g. Acropora cervicornis would be ACR CERV).}
}
\value{
A dataframe or list of dataframes containing all demographic data for specified region, and species, if selected. A list is produced for Florida regions, with one dataframe containing all single stage data and the other containing all two stage data.
}
\description{
Loads combined dataframes of coral demographics data from all years of a single region. Can be just NCRMP
data, or also NCRMP and DRM data combined.
In regions with two stage data (SEFCRI, FLK, Tortugas) a list of dataframes is produces, one
containing all the single stage data, one containing all the two stage data.
This function is called by nearly all of the functions that run summary calculations from coral demographic data.
}
