% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_NTOT.R
\name{load_NTOT}
\alias{load_NTOT}
\title{Creates weighted demo data}
\usage{
load_NTOT(region, inputdata, project)
}
\arguments{
\item{region}{A string indicating the region}

\item{inputdata}{A dataframe}

\item{project}{A string indicating the project, NCRMP or NCRMP and DRM combined}
}
\value{
A dataframe
}
\description{
Creates weighted demo data
}
