% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_NTOT_species.R
\name{load_NTOT_species}
\alias{load_NTOT_species}
\title{Creates weighted demo data, for individual species}
\usage{
load_NTOT_species(region, inputdata, project)
}
\arguments{
\item{region}{A string indicating the region}

\item{inputdata}{A dataframe}

\item{project}{A string indicating the project, NCRMP or NCRMP and DRM combined}
}
\description{
Creates weighted demo data, for individual species
}
