% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_SCREAM_analysis_ready_data.R
\name{make_SCREAM_analysis_ready_data}
\alias{make_SCREAM_analysis_ready_data}
\title{Clean and tidy preNCRMP data}
\usage{
make_SCREAM_analysis_ready_data()
}
\arguments{
\item{region}{A string indicating the region}

\item{userlocation}{A string indicating the user location}
}
\value{
A list of multiple dataframes
}
\description{
Clean and tidy preNCRMP data
}
