% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_size_bins.R
\name{make_size_bins}
\alias{make_size_bins}
\title{Calculates bin sizes for strata and domain estimates of size and length}
\usage{
make_size_bins(
  region,
  project = "NCRMP",
  years,
  size_bin_count = 10,
  length_bin_count = 10,
  species_filter = NULL
)
}
\arguments{
\item{region}{A string indicating the region}

\item{project}{A string indicating the project, NCRMP or NCRMP and DRM combined}

\item{years}{A numeric vector containing the years to be observed}

\item{size_bin_count}{The number of desired bins for length}

\item{species_filter}{A character vector of species codes to be observed (default = NULL i.e. no filter)}
}
\value{
A list of six dataframes
}
\description{
Calculates bin sizes for strata and domain estimates of size and length
}
\examples{
analyzed_species <- c("ACR CERV", "ACR PALM", "ORB ANNU")
make_size_bins(region = "SEFCRI", years = c(2014, 2020), species_filter = analyzed_species) 
}
