% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRMP_calculate_invert_density.R
\name{NCRMP_calculate_invert_density}
\alias{NCRMP_calculate_invert_density}
\title{Creates colony density and colony size summary dataframes}
\usage{
NCRMP_calculate_invert_density(region, project = "NULL")
}
\arguments{
\item{region}{A string indicating the region}

\item{project}{A string indicating the project: "NCRMP" or "MIR". Default is NCRMP.}

\item{inputdf}{A dataframe}
}
\value{
A dataframe
}
\description{
Creates colony density and colony size summary dataframes
}
