% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRMP_colony_density_CV_and_occurrence.R
\name{NCRMP_colony_density_CV_and_occurrence}
\alias{NCRMP_colony_density_CV_and_occurrence}
\title{Creates colony density summary dataframes}
\usage{
NCRMP_colony_density_CV_and_occurrence(
  region,
  ptitle,
  file_path,
  species_filter = "NULL"
)
}
\arguments{
\item{region}{A string indicating the region}

\item{ptitle}{A string indicating the plot title}

\item{path}{A string indicating the filepath for the figure}
}
\value{
A dataframe
}
\description{
Creates colony density summary dataframes
}
