% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRMP_DRM_calculate_colony_density.R
\name{NCRMP_DRM_calculate_colony_density}
\alias{NCRMP_DRM_calculate_colony_density}
\title{Creates colony density summary dataframes}
\usage{
NCRMP_DRM_calculate_colony_density(
  project = "NULL",
  region,
  species_filter = "NULL"
)
}
\arguments{
\item{project}{A string indicating the project, NCRMP or NCRMP and DRM combined}

\item{region}{A string indicating the region}

\item{species_filter}{A string indicating whether to filter to a subset of species}
}
\value{
A dataframe
}
\description{
Creates colony density summary dataframes
}
