% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/NCRMP_DRM_calculate_dis_ble_prevalence_species_domain.R
\name{NCRMP_DRM_calculate_dis_ble_prevalence_species_domain}
\alias{NCRMP_DRM_calculate_dis_ble_prevalence_species_domain}
\title{Species domain estimates for disease prevalence & bleaching prevalence for NCRMP and NCRMP + DRM data}
\usage{
NCRMP_DRM_calculate_dis_ble_prevalence_species_domain(project, region)
}
\arguments{
\item{project}{A string indicating the project, NCRMP, MIR, or NCRMP and DRM combined ("NCRMP_DRM").}

\item{region}{A string indicating the region. Options are: "SEFCRI", "FLK", "Tortugas", "STX", "STTSTJ", "PRICO", and "GOM".}
}
\value{
A list dataframes, 1) bleaching prevalence domain estimates and
2) disease prevalence domain estimates, for all years of a given region.
}
\description{
Calculates regional domain estimates for disease and bleaching prevalence by species,
for a specified region. NCRMP utilizes a stratified random sampling design.
Regional estimates of prevalence are weighted by the number of
grid cells of a stratum in the sample frame.
Uses data summaries created by [NCRMP_DRM_calculate_disease_prevalence_colonies()] function
(disease and bleaching prevalence by species and site).
}
