% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRMP_DRM_calculate_mean_colony_size.R
\name{NCRMP_DRM_calculate_mean_colony_size}
\alias{NCRMP_DRM_calculate_mean_colony_size}
\title{Creates mean colony size summary dataframes}
\usage{
NCRMP_DRM_calculate_mean_colony_size(
  project = "NULL",
  region,
  species_filter = "NULL"
)
}
\arguments{
\item{project}{A string indicating the project, NCRMP or NCRMP and DRM combined}

\item{region}{A string indicating the region}

\item{species_filter}{A string indicating whether to filter to a subset of species}
}
\value{
A dataframe
}
\description{
Creates mean colony size summary dataframes
}
