% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRMP_DRM_calculate_mortality.R
\name{NCRMP_DRM_calculate_mortality}
\alias{NCRMP_DRM_calculate_mortality}
\title{Creates colony mortality summary dataframes}
\usage{
NCRMP_DRM_calculate_mortality(project, region, species_filter = "NULL")
}
\arguments{
\item{project}{A string indicating the project, NCRMP ("NCRMP") or NCRMP and DRM combined ("NCRMO_DRM").}

\item{region}{A string indicating the region. Options include: "FLK", "SEFCRI", "Tortugas", "STX", "STTSTJ", and "GOM".}

\item{species_filter}{An optional string indicating whether to filter to a subset of species.}
}
\value{
A list of dataframes including 1) mean old mortality at each site, 2)
mean recent mortality at each site, 3) mean old mortality in each strata, 4)
mean recent mortality in each strata, 5) mean old mortality for each species in
each strata, 6) mean recent mortality for each species in each strata, 7)
regional estimate for old mortality, 8) regional estimate for recent mortality,
9) regional estimate for old mortality for each species, 10) regional
estimate for recent mortality for each species.
}
\description{
Calculates mean old and recent coral mortality at each site, strata, and
region from the NCRMP or NCRMP+DRM coral demographic data.
Means are also calculated for each species at each strata and for the region.
Regional estimates of mortality are weighted by the number of grid cells
of a stratum in the sample frame.
}
