% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/NCRMP_DRM_calculate_species_richness_diversity.R
\name{NCRMP_DRM_calculate_species_richness_diversity}
\alias{NCRMP_DRM_calculate_species_richness_diversity}
\title{Creates species list, species richness and species diversity dataframes from NCRMP and FRRP data}
\usage{
NCRMP_DRM_calculate_species_richness_diversity(project, region)
}
\arguments{
\item{project}{A string indicating the project, NCRMP or NCRMP and FRRP combined}

\item{region}{A string indicating the region}

\item{analysis_strat}{A string indicating the analysis level strata}
}
\value{
A dataframe
}
\description{
Creates species list, species richness and species diversity dataframes from NCRMP and FRRP data
}
