% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRMP_DRM_colony_density_CV_and_occurrence.R
\name{NCRMP_DRM_colony_density_CV_and_occurrence}
\alias{NCRMP_DRM_colony_density_CV_and_occurrence}
\title{Creates colony density summary dataframes}
\usage{
NCRMP_DRM_colony_density_CV_and_occurrence(
  region,
  ptitle,
  year,
  file_path = "NULL",
  species_filter = "NULL",
  project
)
}
\arguments{
\item{region}{A string indicating the region. Options are: "SEFCRI", "FLK", "Tortugas", "STX", "STTSTJ", "PRICO", and "GOM".}

\item{ptitle}{A string indicating the plot title, usually the region.}

\item{year}{A numeric indicating the year of interest, which will be plotted.}

\item{path}{A string indicating the filepath for the figure.}
}
\value{
A list dataframes. First, a dataframe of regional weighted mean density, CV,
and occurrence, by species for a given region. Second, a dataframe of the same,
filtered to only species/years where CV is less than or equal to 20%.
}
\description{
Calculates regional estimate of coral density and coefficient of variation (CV),
by species, for a given region. NCRMP utilizes a stratified random
sampling design. Regional estimates of density are weighted by the number of grid cells of a stratum
in the sample frame. Function calculates weighted strata means to produce
regional estimates for coral density data by species.
Also calculates occurrence of each species in each year.
Additionally, function produces a figure of species occurrence and CV by species
for a given region and year.
}
