% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRMP_FRRP_calculate_colony_density.R
\name{NCRMP_FRRP_calculate_colony_density}
\alias{NCRMP_FRRP_calculate_colony_density}
\title{Creates colony density summary dataframes}
\usage{
NCRMP_FRRP_calculate_colony_density(project, region,
  species_filter = "NULL", analysis_strat = "NULL")
}
\arguments{
\item{project}{A string indicating the project, NCRMP or NCRMP and FRRP combined}

\item{region}{A string indicating the region}

\item{species_filter}{A string indicating whether to filter to a subset of species}

\item{analysis_strat}{A string indicating the analysis level strata}
}
\value{
A dataframe
}
\description{
Creates colony density summary dataframes
}
