% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRMP_FRRP_calculate_disease_prevalence.R
\name{NCRMP_FRRP_calculate_disease_prevalence}
\alias{NCRMP_FRRP_calculate_disease_prevalence}
\title{Creates species list, species richness and species diversity dataframes from NCRMP and FRRP data}
\usage{
NCRMP_FRRP_calculate_disease_prevalence(project, region,
  species_filter = "NULL")
}
\arguments{
\item{project}{A string indicating the project, NCRMP or NCRMP and FRRP combined}

\item{region}{A string indicating the region}
}
\value{
A dataframe
}
\description{
Creates species list, species richness and species diversity dataframes from NCRMP and FRRP data
}
