% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRMP_make_bleaching_abundance.R
\name{NCRMP_make_bleaching_abundance}
\alias{NCRMP_make_bleaching_abundance}
\title{Creates abundance per bleaching category dataframe for all regions and all years}
\usage{
NCRMP_make_bleaching_abundance()
}
\value{
A dataframe
}
\description{
Creates abundance per bleaching category dataframe for all regions and all years
}
