% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRMP_make_size_bins.R
\name{NCRMP_make_size_bins}
\alias{NCRMP_make_size_bins}
\title{Outputs the strata and domain estimates for 3D surface area and length}
\usage{
NCRMP_make_size_bins(
  region,
  project = "NCRMP",
  years,
  size_bin_count = 10,
  length_bin_count = 10,
  species_filter = NULL
)
}
\arguments{
\item{region}{A string indicating the region}

\item{project}{A string indicating the project, NCRMP or NCRMP and DRM combined}

\item{years}{A string indicating the two years to compare. Must be NCRMP sampling years.}

\item{size_bin_count}{A number indicating the desired number of bins for 3d surface area}

\item{length_bin_count}{A number indicating the desired number of bins for length}

\item{species_filter}{A string indicating whether to filter to a subset of species}
}
\value{
A dataframe
}
\description{
Outputs the strata and domain estimates for 3D surface area and length
}
