% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRMP_make_weighted_invert_density_data.R
\name{NCRMP_make_weighted_invert_density_data}
\alias{NCRMP_make_weighted_invert_density_data}
\title{Creates weighted Invert density data}
\usage{
NCRMP_make_weighted_invert_density_data(inputdata, region)
}
\arguments{
\item{inputdata}{A dataframe}

\item{region}{A string indicating the region}
}
\value{
A dataframe
}
\description{
Creates weighted Invert density data
}
