% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRMP_make_weighted_LPI_data.R
\name{NCRMP_make_weighted_LPI_data}
\alias{NCRMP_make_weighted_LPI_data}
\title{Creates weighted LPI data}
\usage{
NCRMP_make_weighted_LPI_data(inputdata, region)
}
\arguments{
\item{inputdata}{A dataframe}

\item{region}{A string indicating the region}
}
\value{
A dataframe
}
\description{
Creates weighted LPI data
}
