% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRMP_make_weighted_LPI_data_RC.R
\name{NCRMP_make_weighted_LPI_data_RC}
\alias{NCRMP_make_weighted_LPI_data_RC}
\title{Creates weighted LPI data}
\usage{
NCRMP_make_weighted_LPI_data_RC(inputdata, region, year)
}
\arguments{
\item{inputdata}{A dataframe}

\item{region}{A string indicating the region}

\item{year}{A string indicating the year}
}
\value{
A dataframe
}
\description{
Creates weighted LPI data
}
