% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRMP_perform_ttest.R
\name{NCRMP_perform_ttest}
\alias{NCRMP_perform_ttest}
\title{Performs t-test comparing estimates}
\usage{
NCRMP_perform_ttest(
  dataframe,
  metric1,
  metric2,
  alpha,
  n_years,
  return_dataframe
)
}
\arguments{
\item{dataframe}{A dataframe of yearly domain estimates for a single subjurisdiction.}

\item{metric1}{The metric you are testing: density, hard coral cover, macroalgae cover, disease prevalence, bleaching prevalence.}

\item{metric2}{The column name of the metric you are testing: avCvr, avDen, etc.}

\item{alpha}{The alpha level you want to test, ex. 0.05}

\item{n_years}{You must specify the number of sampling years completed in the subjurisdiction if test_type = "Years", currently 4 for St. Thomas & St. John and 3 for all other regions.}

\item{test_type}{You can compare domain estimates between years or inside/ outside a protected area for a single year, enter "Years" or "PROT"}
}
\value{
A dataframe and test results as a list.
}
\description{
Function to perform t-test on various summaries of NCRMP data.
Typically, comparisons are made between yearly domain estimates (e.g. mean
coral cover by year) in a single region. But comparisons could be made between
inside/outside a protected area.
}
