% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_2stage_data.R
\name{subset_2stage_data}
\alias{subset_2stage_data}
\title{Subset 2 stage data}
\usage{
subset_2stage_data(region, project, year)
}
\arguments{
\item{region}{A string indicating the region}

\item{project}{A string indicating the project, NCRMP or NCRMP and DRM combined}

\item{year}{A numeric indicating the year of interest}
}
\value{
# A subsetted dataframe
}
\description{
Subsets 2 stage data to single stage to provide data frames that can be used
for the size frequency data summaries. This function should NOT be called and run
but serves as base code for randomly subsetting 2 stage data into single stage
and saving it to the package.
}
