% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_NCRMP_DRM_demo_data.R
\name{load_NCRMP_DRM_demo_data}
\alias{load_NCRMP_DRM_demo_data}
\title{Creates combined demo dataframes}
\usage{
load_NCRMP_DRM_demo_data(project = "NULL", region, species_filter = "NULL")
}
\arguments{
\item{project}{A string indicating the project, NCRMP or NCRMP and DRM combined}

\item{region}{A string indicating the region}

\item{species_filter}{A string indicating whether to filter to a subset of species}
}
\value{
A dataframe
}
\description{
Creates combined demo dataframes
}
