% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cover_darwincore.R
\name{make_cover_darwincore}
\alias{make_cover_darwincore}
\title{Format analysis ready LPI data into Darwin Core}
\usage{
make_cover_darwincore(dat, region, year)
}
\arguments{
\item{dat}{A dataframe of analysis ready LPI data to be converted into Darwin Core}

\item{region}{A string indicating the region}

\item{year}{A string indicating the year}
}
\value{
A list of multiple dataframes including formatted data
}
\description{
Format analysis ready LPI data into Darwin Core
}
