% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_demo_darwincore.R
\name{make_demo_darwincore}
\alias{make_demo_darwincore}
\title{Format analysis ready demo data into Darwin Core}
\usage{
make_demo_darwincore(dat)
}
\arguments{
\item{dat}{A dataframe of analysis ready coral demographic data to be converted into Darwin Core format.}
}
\value{
A list of multiple dataframes including re-formatted data and another that checks for duplicate occurrence ID's.
}
\description{
Format analysis ready demo data into Darwin Core
}
