% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRMP_calculate_cover.R
\name{NCRMP_calculate_cover}
\alias{NCRMP_calculate_cover}
\title{Creates percent cover dataframe}
\usage{
NCRMP_calculate_cover(region, year, analysis_strat = "NULL")
}
\arguments{
\item{region}{A string indicating the region}

\item{year}{A string indicating the year}

\item{analysis_strat}{A string indicating the analysis level strata}
}
\value{
A dataframe
}
\description{
Creates percent cover dataframe
}
