% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRMP_calculate_cover.R
\name{NCRMP_calculate_cover}
\alias{NCRMP_calculate_cover}
\title{Creates percent cover dataframe}
\usage{
NCRMP_calculate_cover(region, project = "NULL")
}
\arguments{
\item{region}{A string indicating the region}

\item{project}{A string indicating the project: "NCRMP" or "MIR". Default is NCRMP.}
}
\value{
A dataframe
}
\description{
Creates percent cover dataframe
}
