% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRMP_calculate_ESA_corals_PresAbs.R
\name{NCRMP_calculate_ESA_corals_PresAbs}
\alias{NCRMP_calculate_ESA_corals_PresAbs}
\title{Creates colony density and colony size summary dataframes}
\usage{
NCRMP_calculate_ESA_corals_PresAbs()
}
\value{
A dataframe
}
\description{
Creates colony density and colony size summary dataframes
}
