% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRMP_calculate_invert_density.R
\name{NCRMP_calculate_invert_density}
\alias{NCRMP_calculate_invert_density}
\title{Creates summary dataframes of Diadema density from NCRMP benthic assessment.
NCRMP utilizes a stratified random sampling design.
Regional estimates of Diadema density are weighted by the number of
grid cells of a stratum in the sample frame.}
\usage{
NCRMP_calculate_invert_density(region, project = "NULL")
}
\arguments{
\item{region}{A string indicating the region. Options are: "SEFCRI", "FLK", "Tortugas", "STX", "STTSTJ", "PRICO", and "GOM".}

\item{project}{A string indicating the project: "NCRMP" or "MIR". Default is NCRMP.}
}
\value{
A list dataframes including 1) Diadema density at each site,
2) mean Diadema density by strata, and 3) weighted regional mean Diadema density,
all for a given region.
}
\description{
Creates summary dataframes of Diadema density from NCRMP benthic assessment.
NCRMP utilizes a stratified random sampling design.
Regional estimates of Diadema density are weighted by the number of
grid cells of a stratum in the sample frame.
}
