% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRMP_colony_density_CV_and_occurrence.R,
%   R/NCRMP_colony_density_CV_and_occurrence_DEPRECATED.R
\name{NCRMP_colony_density_CV_and_occurrence}
\alias{NCRMP_colony_density_CV_and_occurrence}
\title{Creates colony density summary dataframes}
\usage{
NCRMP_colony_density_CV_and_occurrence(
  region,
  ptitle,
  year,
  file_path = "NULL",
  species_filter = "NULL",
  project = "NULL"
)

NCRMP_colony_density_CV_and_occurrence(
  region,
  ptitle,
  year,
  file_path = "NULL",
  species_filter = "NULL",
  project = "NULL"
)
}
\arguments{
\item{region}{A string indicating the region.  Options are: "SEFCRI", "FLK", "Tortugas", "STX", "STTSTJ", "PRICO", and "GOM".}

\item{ptitle}{A string indicating the plot title.}

\item{year}{A numeric indicating the year of interest. Only controls which data are plotted.}

\item{project}{A sting indicating the project. Only option is "NCRMP".}

\item{path}{A string indicating the filepath for the figure}
}
\value{
A dataframe

A list of dataframes, including one with all regional weighted mean densities,
CV's, and occurrences per species per year, and another with the same but filtered
to species/years where CV <= 20%
}
\description{
DEPRECATED. Use [NCRMP_DRM_colony_density_CV_and_occurrence()] instead.
}
\details{
Calculates regional weighted mean coral density and coefficient of variation (CV), by species,
from coral demographic data. NCRMP utilizes a stratified random
sampling design. Regional estimates of density are weighted by the number of grid cells of a stratum
in the sample frame. Function calculates strata means, weighted strata means,
and weighted regional estimates for various measures from coral demographic data.
Also calculates occurrence of each species in each year. Produces a figure
of occurrence and CV by species for specified year.
Utilizes [NCRMP_make_weighted_density_CV_data()] to calculate regional
mean density, CV, and occurrence.
}
