% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRMP_colony_percent cover.R
\name{NCRMP_colony_percent_cover}
\alias{NCRMP_colony_percent_cover}
\title{Creates colony coral cover summary dataframes}
\usage{
NCRMP_colony_percent_cover(region, ptitle, file_path)
}
\arguments{
\item{region}{A string indicating the region}

\item{ptitle}{A string indicating the plot title}

\item{path}{A string indicating the filepath for the figure}
}
\value{
A dataframe
}
\description{
Creates colony coral cover summary dataframes
}
