% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRMP_colony_percent cover.R
\name{NCRMP_colony_percent_cover}
\alias{NCRMP_colony_percent_cover}
\title{Creates colony coral cover summary data frames}
\usage{
NCRMP_colony_percent_cover(region, ptitle, file_path, project = "NULL")
}
\arguments{
\item{region}{A string indicating the region. Options are: "SEFCRI", "FLK", "Tortugas", "STX", "STTSTJ", "PRICO", and "GOM".}

\item{ptitle}{A string indicating the plot title.}

\item{file_path}{A string indicating the filepath for the figure output by this function.}

\item{project}{A string indicating the project: "NCRMP" or "MIR".}
}
\value{
A list of dataframes and a jpeg. Dataframes include strata mean cover by species and regional weighted mean cover by species.
}
\description{
Calculates coral cover strata means and regional weighted means, by coral species. Also
produces a figure of percent cover by species from most recent year of data for the
selected region.
}
