% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/NCRMP_DRM_calculate_dis_ble_prevalence_species_domain.R
\name{NCRMP_DRM_calculate_dis_ble_prevalence_species_domain}
\alias{NCRMP_DRM_calculate_dis_ble_prevalence_species_domain}
\title{Species domain estimates for disease prevalence & bleaching prevalence for NCRMP and NCRMP + DRM data}
\usage{
NCRMP_DRM_calculate_dis_ble_prevalence_species_domain(project, region)
}
\arguments{
\item{project}{A string indicating the project, NCRMP or NCRMP and DRM combined}

\item{region}{A string indicating the region}
}
\value{
A dataframe
}
\description{
Species domain estimates for disease prevalence & bleaching prevalence for NCRMP and NCRMP + DRM data
}
