% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/NCRMP_DRM_calculate_disease_prevalence_colonies.R
\name{NCRMP_DRM_calculate_disease_prevalence_colonies}
\alias{NCRMP_DRM_calculate_disease_prevalence_colonies}
\title{disease prevalence & bleaching prevalence for NCRMP and NCRMP + DRM data at the species/site, site, strata and domain levels}
\usage{
NCRMP_DRM_calculate_disease_prevalence_colonies(
  project,
  region,
  species_filter = "NULL"
)
}
\arguments{
\item{project}{A string indicating the project, NCRMP, MIR, or NCRMP and DRM combined}

\item{region}{A string indicating the region}
}
\value{
A dataframe
}
\description{
disease prevalence & bleaching prevalence for NCRMP and NCRMP + DRM data at the species/site, site, strata and domain levels
}
