% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRMP_DRM_calculate_disease_prevalence_sites.R
\name{NCRMP_DRM_calculate_disease_prevalence_sites}
\alias{NCRMP_DRM_calculate_disease_prevalence_sites}
\title{Creates species list, species richness and species diversity dataframes from NCRMP and DRM data}
\usage{
NCRMP_DRM_calculate_disease_prevalence_sites(
  project,
  region,
  species_filter = "NULL"
)
}
\arguments{
\item{project}{A string indicating the project, NCRMP or NCRMP and DRM combined}

\item{region}{A string indicating the region}
}
\value{
A dataframe
}
\description{
Creates species list, species richness and species diversity dataframes from NCRMP and DRM data
}
