% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRMP_DRM_calculate_mortality.R
\name{NCRMP_DRM_calculate_mortality}
\alias{NCRMP_DRM_calculate_mortality}
\title{Creates colony mortality summary dataframes}
\usage{
NCRMP_DRM_calculate_mortality(project, region, species_filter = "NULL")
}
\arguments{
\item{project}{A string indicating the project, NCRMP ("NCRMP") or NCRMP and DRM combined ("NCRMO_DRM")}

\item{region}{A string indicating the region. Options include: "FLK", "SEFCRI", "Tortugas", "STX", "STTSTJ", and "GOM"}

\item{species_filter}{A boolean value indicating whether to filter to a subset of species}
}
\value{
A dataframe
}
\description{
Creates colony mortality summary dataframes
}
