% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/NCRMP_DRM_calculate_species_richness_diversity.R
\name{NCRMP_DRM_calculate_species_richness_diversity}
\alias{NCRMP_DRM_calculate_species_richness_diversity}
\title{Creates species list, species richness and species diversity dataframes from NCRMP and DRM data}
\usage{
NCRMP_DRM_calculate_species_richness_diversity(project, region)
}
\arguments{
\item{project}{A string indicating the project, NCRMP ("NCRMP") or NCRMP and DRM combined ("NCRMP_DRM").}

\item{region}{A string indicating the region. Options are: "SEFCRI", "FLK", "Tortugas", "STTSTJ", "STX, "GOM" or "PRICO".}
}
\value{
A list dataframes
}
\description{
Creates data summaries of species richness and diversity, based on NCRMP
coral demographic data.
Species richness includes juveniles. Diversity is based only on adults.
Note that there is no accounting for transect length at this point, for example,
we do not calculate # of species/m^2. However, transect length may vary.
For 2-stage data (Florida, NCRMP+DRM), richness and diversity are averaged
between the 2 stations. No coral sites are not included in these estimates of
richness and no adult coral sites are not included in the estimates of diversity.
}
