% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRMP_make_size_bins.R
\name{NCRMP_make_size_bins}
\alias{NCRMP_make_size_bins}
\title{Outputs the strata and domain estimates for 3D surface area and length}
\usage{
NCRMP_make_size_bins(
  region,
  project,
  years,
  size_bin_count = 10,
  length_bin_count = 10,
  species_filter = NULL
)
}
\arguments{
\item{region}{A string indicating the region.  Options are: "SEFCRI", "FLK", "Tortugas", "STX", "STTSTJ", "PRICO", and "GOM".}

\item{project}{A string indicating the project, NCRMP or NCRMP and DRM combined ("NCRMP_DRM").}

\item{years}{A concatenation of numerics indicating the two years to compare. Must be NCRMP sampling years if project = "NCRMP".}

\item{size_bin_count}{A number indicating the desired number of bins for 3d surface area.
As function is currently set up, this is not used.}

\item{length_bin_count}{A number indicating the desired number of bins for length.
As function is currently set up, this is not used.}

\item{species_filter}{A concatenated string indicating whether to filter to a subset of species}
}
\value{
A list of dataframes.
}
\description{
Creates regional weighted densities at size (3D surface and length),
as well as regional weighted relative length frequencies from NCRMP coral
demographic data. NCRMP utilizes a stratified random
sampling design. Regional estimates are weighted by the number of
grid cells of a stratum in the sample frame. For coral size data, weighting
is done only by strata where species is present.
}
