% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRMP_make_weighted_demo_data.R
\name{NCRMP_make_weighted_demo_data}
\alias{NCRMP_make_weighted_demo_data}
\title{Creates weighted demo data}
\usage{
NCRMP_make_weighted_demo_data(
  project,
  inputdata,
  region,
  datatype,
  species_filter,
  species_data = NULL
)
}
\arguments{
\item{project}{A string indicating the project, NCRMP or NCRMP and DRM combined}

\item{inputdata}{A dataframe}

\item{region}{A string indicating the region}

\item{datatype}{A string indicating the datatype}
}
\value{
A dataframe
}
\description{
Creates weighted demo data
}
