% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRMP_make_weighted_species_coral_cover_data.R
\name{NCRMP_make_weighted_species_coral_cover_data}
\alias{NCRMP_make_weighted_species_coral_cover_data}
\title{Calculate weighted mean density & CVs}
\usage{
NCRMP_make_weighted_species_coral_cover_data(region, sppcvr)
}
\arguments{
\item{region}{A string indicating the region}
}
\value{
A dataframe
}
\description{
Calculate weighted mean density & CVs
}
