% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCRMP_perform_ttest.R
\name{NCRMP_perform_ttest}
\alias{NCRMP_perform_ttest}
\title{Creates percent cover dataframe}
\usage{
NCRMP_perform_ttest(
  dataframe,
  metric1,
  metric2,
  alpha,
  n_years,
  return_dataframe
)
}
\arguments{
\item{dataframe}{A dataframe of yearly domain estimates for a single subjurisdiction}

\item{metric1}{The metric you are testing: density, hard coral cover, macroalgae cover, disease prevalence, bleaching prevalence}

\item{metric2}{The column name of the metric you are testing: avCvr, avDen, etc.}

\item{alpha}{THe alpha level you want to test, ex. 0.05}

\item{n_years}{The number of sampling years completed in the subjurisdiction, currently 4 for St. Thomas & St. John and 3 for all other regions}
}
\value{
A dataframe
}
\description{
Creates percent cover dataframe
}
