% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_2stage_data.R
\name{subset_2stage_data}
\alias{subset_2stage_data}
\title{Title}
\usage{
subset_2stage_data(region, project, year)
}
\arguments{
\item{region}{A string indicating the region}

\item{project}{A string indicating the project, NCRMP or NCRMP and DRM combined}

\item{year}{A numeric indicating the year of interest}
}
\value{
# A subsetted dataframe
}
\description{
Title
}
